% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-fields.R
\name{check_field_requirements}
\alias{check_field_requirements}
\title{Check field requirements}
\usage{
check_field_requirements(field, dic)
}
\arguments{
\item{field}{Field requirements}

\item{dic}{Extended dictionary}
}
\value{
List of checked field requirements
}
\description{
Create list of validated field requirements (before checking against OTHER field requirements)
}
