% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{create_dic}
\alias{create_dic}
\title{Create advanced dictionary}
\usage{
create_dic(d, frtype = NULL, extended = FALSE)
}
\arguments{
\item{df}{Dataframe or tibble for which to create dictionary}
}
\value{
Dataframe with dictionary for df
}
\description{
TODO integrate into homodatum's create_dic!!!!
}
\details{
This is an extended/advanced version of the dicionary created by \code{homodatum::create_dic}.
Further down the line, this will be moved to \code{homodatum}, potentially integrated in the
existing dictionary function by \code{homodatum::create_dic(..., extended = TRUE)}.

Additionally to the existing id, label, and hdType this should have
\itemize{
\item the base R class of the variable (numeric, character, ...)
\item number of distinct categories (for Cat, NULL otherwise)
\item min, max, mean, median values (for Num, NULL otherwise)
\item number format, i.e. decimal separator (for Num, NULL otherwise)
\item number of missing values
\item anything else?
}
}
