% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{create_dic_advanced}
\alias{create_dic_advanced}
\title{Create advanced dictionary}
\usage{
create_dic_advanced(df)
}
\arguments{
\item{df}{}
}
\value{
Dataframe with dictionary for df
}
\description{
An extended/advanced version of the dicionary created by homodatum::create_dic
This could be part of homodatum, potentially integrated in the existing dictionary
function by homodatum::create_dic(..., extended = TRUE).
}
\details{
Additionally to the existing id, label, and hdType this should have
\itemize{
\item the base R class of the variable (numeric, character, ...)
\item number of distinct categories (for Cat, NULL otherwise)
\item min, max, mean, median values (for Num, NULL otherwise)
\item number format, i.e. decimal separator (for Num, NULL otherwise)
\item number of missing values
\item anything else?
}
}
