% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-fields.R
\name{order_fields}
\alias{order_fields}
\title{Order fields}
\usage{
order_fields(fields)
}
\arguments{
\item{fields}{List of field specifications}
}
\value{
Ordered list of field specification
}
\description{
Order field specifications by diff_want_is.
}
