% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{recommend_app}
\alias{recommend_app}
\title{Recommends suitable app}
\usage{
recommend_app(df)
}
\arguments{
\item{df}{Loaded dataframe}
}
\value{
List of recommended DS apps
}
\description{
ADVANCED
}
\details{
Suggest available DS apps from a loaded df.

This uses recommend_viz to suggest the DS apps that can be used
to create recommended visualisations from the loaded df.
}
