% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend.R
\name{recommend_viz}
\alias{recommend_viz}
\title{Recommends suitable visualization}
\usage{
recommend_viz(df)
}
\arguments{
\item{df}{Loaded dataframe}
}
\value{
List of recommended visualizations
}
\description{
ADVANCED
}
\details{
Suggest suitable visualization types from a loaded df.

For this, we would need a list of all possible visualisation we offer
with DS apps and sets of minimum requirements for each of them.

e.g. app-sankey-interactive: 2+ columns with categorical variables with max 20 categories (more?)
app-simple-charts-orgs: 2+ columns in total AND need at least one categorical column (more?)
}
