% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{requirements_load}
\alias{requirements_load}
\title{Load requirements for validation}
\usage{
requirements_load(path = "dsvalidate", filename = "requirements.yaml")
}
\arguments{
\item{path}{Path for the requirements.yaml file; default is 'dsvalidate'}

\item{filename}{Filename of requirements file, default is 'requirements.yaml'}
}
\value{
List of requirements
}
\description{
Loads and checks the requirements.yaml file, raises necessary
errors and warnings if necessary, and prepares it for use by
validate_requirements.
}
