% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-fields.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate columns}
\usage{
validate_columns(dic, specs, field_id)
}
\arguments{
\item{dic}{Extended dictionary, created by created_dic(df, extended = TRUE)}

\item{specs}{List of field specifications}

\item{field_id}{Field id}
}
\value{
Tibble with an 'id' column for the column id, a boolean 'meets_requirements' column indicating whether the column meets
the field specifications and a boolean 'matches_id' column indicating whether the column id is equal to the field id.
}
\description{
For all field specifications, check for each column in dic whether or not the column meets the fields requirements.
}
