% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-fields.R
\name{validate_fields}
\alias{validate_fields}
\title{Validate fields}
\usage{
validate_fields(
  x,
  requirements,
  validated_table_meta = TRUE,
  validated_table_specs = TRUE
)
}
\arguments{
\item{x}{Dataframe or list of dataframe}

\item{requirements}{List of requirements}

\item{validated_table_meta}{Boolean, whether or not table meta requirements were met}

\item{validated_table_specs}{Boolean, whether or not table specs requirements were met}
}
\value{
List of validations
}
\description{
For each table, validates all field specifications
}
