% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-table-meta.R
\name{validate_table_meta}
\alias{validate_table_meta}
\title{Validate table meta data}
\usage{
validate_table_meta(x, requirements)
}
\arguments{
\item{x}{Dataframe or list of dataframe}

\item{requirements}{List of requirements}
}
\value{
List of validations
}
\description{
Validates input data for
}
\details{
\itemize{
\item data type
\item number of tables
\item table ids (names)
}
}
