% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-table-specs.R
\name{validate_table_specs}
\alias{validate_table_specs}
\title{Validate table specifications}
\usage{
validate_table_specs(x, requirements, validated_table_meta = TRUE)
}
\arguments{
\item{x}{Dataframe or list of dataframe}

\item{requirements}{List of requirements}
}
\value{
List of validations
}
\description{
Validates each table specification by checking tables for
}
\details{
\itemize{
\item number of columns
\item number of rows
\item frTypes
}
}
