% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{requirements_load}
\alias{requirements_load}
\title{Load requirements for validation}
\usage{
requirements_load(path = "dsvalidate")
}
\arguments{
\item{path}{Path for the requirements.yaml file; default is 'dsvalidate'}
}
\value{
List of requirements
}
\description{
Loads and checks the requirements.yaml file, raises necessary
errors and warnings if necessary, and prepares it for use by
validate_requirements.
}
\details{
The list obtained from the requirements.yaml file is a list with a sublist
for each requirement. Sublists (requirements are connected with logical AND).

TODO: Would be good to enable AND and OR logic for requirements
}
\examples{
# Example 1: Need 2+ columns in total AND need at least one categorical column.
# The logic behind the requirements in the yaml file is:
# - requirement 1: number of columns needs to be greater than 1
# - requirement 2: number of columns where hdType is equal to Cat needs to be greater than 0
req1 <- yaml::read_yaml("inst/test_dsvalidate/ex01-two-columns-one-categorical/dsvalidate/requirements.yaml")
req1

# Example 2: Need 2+ columns in total with fewer than 20 categories each
# The logic behind the requirements in the yaml file is:
# - number of columns where hdType is equal to Cat and n_distinct is less than 20 needs to be greater than 1
req2 <- yaml::read_yaml("inst/test_dsvalidate/ex02-two-categorical-cols-max-distinct-cat/dsvalidate/requirements.yaml")
req2
}
