% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_requirements}
\alias{validate_requirements}
\title{Validate dataframe for requirements}
\usage{
validate_requirements(df, requirements)
}
\arguments{
\item{df}{Dataframe to be validated}

\item{requirements}{List of requirements to be validated}
}
\value{
List object which returns for each requirement whether it df satisfies it,
using label param passed as part of requirements
}
\description{
Validates whether a given df has the necessary requirements.
}
