% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_requirements}
\alias{validate_requirements}
\title{Validate dataframe for requirements}
\usage{
validate_requirements(x, requirements)
}
\arguments{
\item{x}{Dataframe or list of dataframes to be validated}

\item{requirements}{List of requirements to be validated}
}
\value{
List object which returns for each requirement whether it x satisfies it,
using label param passed as part of requirements
}
\description{
Validates whether a given dataframe of list of dataframes meets the requirements
specified in \code{requirements}
}
